Use double instead of int

+++ Data_Buffer/data_buffer.c
@@ -7,7 +7,7 @@ Buffer CreateNewBuffer(int bufferSize)
 {
     Buffer buffer;
     
-    int array[bufferSize];
+    double array[bufferSize];
     
     buffer.queue = array;
     buffer.capacity = bufferSize;
@@ -18,7 +18,7 @@ Buffer CreateNewBuffer(int bufferSize)
 
 
 //this function adds a new data point to the end of the buffer queue, removing the first data point in the queue if the buffer is full
-void Enqueue(Buffer* b, int dataPoint)
+void Enqueue(Buffer* b, double dataPoint)
 {
     if (b->elementsInQueue == b->capacity)
     {
@@ -40,12 +40,12 @@ void Enqueue(Buffer* b, int dataPoint)
 
 
 //this function returns the first data point in the buffer and then removes it from the buffer, assuming the buffer is not empty
-//returns the max value for int if the buffer is empty as an error indicator
-int Dequeue(Buffer* b)
+//returns -1 if the buffer is empty as an error indicator
+double Dequeue(Buffer* b)
 {
     if (b->elementsInQueue > 0)
     {
-        int firstElement = b->queue[0];
+        double firstElement = b->queue[0];
 
         for (int i = 1; i < b->elementsInQueue; i++)
         {
